Remove notion of transaction
Replace transaction with error/request/response.
